Extract some repeated test client fixtures to be used everywhere
